+++ src/js/core/Config.js
@@ -6,7 +6,7 @@ const Config = {
 
     appName: "TrailBook",
 
-    version: "0.7.0",
+    version: "0.8.0",
 
     map: {